Change the query format and redo the query resolution code
At the core of urungi is the query object, which is built
in the report editor and used to fetch the data to be displayed.
There were until now multiple issues with the way queries worked :
* The query object was different then the report.properties object,
	and needed to be synchronised with it, resulting in heavy and
	error prone code for report modification
* The query object was built using layer information and asked directly for
	database columns instead of asking for layer elements
	which means that the code did not follow the logic of the UI,
	which is that the layers is the interface between the report and the database
* The query resolution functions contained absolutely no database security whatsoever,
	allowing anyone who could view the report to use SQL injection and
	recover data at will
* The query resoltion code was clumsy and difficult to understand
* The table join clause was built with a rough, unreliable algorithm which would not work
	for some complicated cases and would create inneficient queries
* Element IDs were handled in a confusing, unreliable way
* The query object was unnecessarely heavy, and it's structure was very complicated,
	which means that in order to change the report edition one had to first know
	how the queries were resolved server side

In light of all these issues, the decision was taken to Redo the query reolution entirely.
This commit creates two new files called queryProcessor.js and new_sql.js.
new_sql.js will replace sql.js in a later branch, and mosty does the same thing, but simpler, and clearer
queryProcessor.js is where most of the changes are. It recieves a very simple query from the client,
and compiles it into a processedQuery which can then be used to create an sql or mongoDb query.
It also guarantees database security.

This commit contains the core of the changes, but further changes will complete this change in organisation
Layers will be modified to do better their role as an interface between database and report
More security will be added
Backend organisation will be improved
MongoDB queries are now broken, and need to be fixed