Use knex.js to resolve queries
Up to this point, sql queries were executed by constructing a sql string with string addition.
This method lead to clumsy security, long and annoying syntax and possible maintainability issues
This commit adds a new Library called knexjs, which gives a functional javascript interface
for sql query building.
For more information, see https://knexjs.org/

knex.js does not support jdbc-oracle and bigquery, so both of these database types are now supported
with the former code, which is located in the legacy folder.