Change SQL query builder to make aggregation work for order and filters
It seems that the way knex was used to aggregate fields in the use-knex-for-query commit was not correct,
and this somehow slipped under the radar.

This commit fixes this, but does so in a rather messy and improvised way.
Although the current solution should not cause any issues,
taking the time to think of a proper way to do it would be preferable.

Unfortunately, there is right now no time to do things well and this quickfix will have to be enough for now.